requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backoff==1.10.0
black==22.3.0
bleach==5.0.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
cisco-opentelemetry-specifications==0.0.21
cleo==0.8.1
click==8.1.3
click-log==0.4.0
colorama==0.4.4
coverage==6.4.1
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.4
docutils==0.18.1
dotty-dict==1.3.0
execnet==1.9.0
executing==0.8.3
fastapi==0.78.0
filelock==3.7.0
Flask==2.1.2
frozenlist==1.3.0
gitdb==4.0.9
GitPython==3.1.27
googleapis-common-protos==1.56.2
grpcio==1.46.3
h11==0.12.0
httpcore==0.15.0
idna==3.3
importlib-metadata==4.11.4
iniconfig==1.1.1
invoke==1.7.1
ipdb==0.13.9
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
keyring==23.5.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
multidict==6.0.2
mypy-extensions==0.4.3
opentelemetry-api==1.10.0
opentelemetry-distro==0.29b0
opentelemetry-exporter-otlp==1.10.0
opentelemetry-exporter-otlp-proto-grpc==1.10.0
opentelemetry-exporter-otlp-proto-http==1.10.0
opentelemetry-instrumentation==0.29b0
opentelemetry-instrumentation-aiohttp-client==0.29b0
opentelemetry-instrumentation-asgi==0.29b0
opentelemetry-instrumentation-fastapi==0.29b0
opentelemetry-instrumentation-flask==0.29b1
opentelemetry-instrumentation-grpc==0.29b0
opentelemetry-instrumentation-requests==0.29b0
opentelemetry-instrumentation-wsgi==0.29b0
opentelemetry-instrumentation-pymongo==0.29b0
opentelemetry-proto==1.10.0
opentelemetry-sdk==1.10.0
opentelemetry-semantic-conventions==0.29b0
opentelemetry-test-utils==0.29b0
opentelemetry-util-http==0.29b0
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
pip==22.0.4
pkginfo==1.8.2
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.29
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
pymongo==4.1.1
pytest==7.1.2
pytest-cov==3.0.0
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-gitlab==3.5.0
python-semantic-release==7.29.0
readme-renderer==35.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
semver==2.13.0
setuptools==58.1.0
setuptools-scm==6.4.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
stack-data==0.2.0
starlette==0.19.1
toml==0.10.2
toml-sort==0.20.0
tomli==2.0.1
tomlkit==0.10.2
tqdm==4.64.0
traitlets==5.2.1.post0
twine==3.8.0
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
